Dodano pierwszy skrypt migracyjny Flyway (V1__init.sql) oraz poprawienie docker-compose.yml application.properties i dodanie zaleznosci w pom.xml aby wszystko dzialalo poprawnie

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,18 +1,18 @@
 CREATE TABLE IF NOT EXISTS role (
-                                    id SERIAL PRIMARY KEY,
+                                    id BIGSERIAL PRIMARY KEY,
                                     name VARCHAR(50) NOT NULL UNIQUE
     );
 
 CREATE TABLE IF NOT EXISTS app_user (
-                                        id SERIAL PRIMARY KEY,
+                                        id BIGSERIAL PRIMARY KEY,
                                         username VARCHAR(100) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    role_id INTEGER REFERENCES role(id)
+    role_id BIGINT REFERENCES role(id)
     );
 
 CREATE TABLE IF NOT EXISTS note (
-                                    id SERIAL PRIMARY KEY,
+                                    id BIGSERIAL PRIMARY KEY,
                                     content TEXT NOT NULL,
                                     created_at TIMESTAMP DEFAULT now(),
-    user_id INTEGER REFERENCES app_user(id) ON DELETE CASCADE
+    user_id BIGINT REFERENCES app_user(id) ON DELETE CASCADE
     );
